api-service/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
httpx==0.24.0
python-dotenv==1.0.0